✨ feat: 더보기 클릭 시 스크롤 이동 구현
