+++ src/app/components/index.ts
@@ -1,5 +1,5 @@
 export { default as ContentsSlider } from './ContentsSlider';
 export { default as EditProfileName } from './EditProfileName';
-export { default as Header } from './Header';
 export { default as PopularFlowers } from './PopularFlowers';
+export { default as ProfileMenuButton } from './ProfileMenuButton';
 export { default as Sidebar } from './Sidebar';
